test: Test that signatures are signed with the expected private key
Uses ssh-keygen to check that tkey-ssh-agent signs using the expected
private key

A context manager is added to TKeySshAgent to let the tkey-ssh-agent run
in the background while calling ssh-keygen.

A custom marker `touch` is added to mark tests as requiring touch

test_uss.py renamed to test_keypair.py
